+++ molo/forms/models.py
@@ -63,8 +63,10 @@ from .widgets import NaturalDateInput
 SKIP = 'NA (Skipped)'
 
 
+ArticlePage.api_fields += ['forms']
 SectionPage.subpage_types += ['forms.MoloFormPage']
 ArticlePage.subpage_types += ['forms.MoloFormPage']
+ArticlePage.content_panels += [InlinePanel('forms', label="Surveys")]
 FooterPage.parent_page_types += ['forms.FormsTermsAndConditionsIndexPage']
 
 
@@ -231,6 +233,9 @@ class MoloFormPage(
                 [FieldPanel('extra_style_hints')], classname="label-above")],
             "Meta")
     ]
+
+    api_fields = ["live", "form_fields", "introduction", "image",
+                  "thank_you_text", "allow_anonymous_submissions"]
 
     def get_effective_extra_style_hints(self):
         return self.extra_style_hints
@@ -798,3 +803,17 @@ class FormsSegmentUserGroup(models.Model):
 
     def __str__(self):
         return self.name
+
+
+class ArticlePageForms(Orderable):
+    page = ParentalKey(ArticlePage, related_name='forms')
+    form = models.ForeignKey(
+        'wagtailcore.Page',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        help_text=_('Survey')
+    )
+    panels = [PageChooserPanel('form', 'forms.MoloFormPage')]
+    api_fields = ['forms']